requirements.txt
# Core dependencies for Bedrock AgentCore and Strands
# Note: Package names may vary. Check official AWS documentation for latest versions.
# The bedrock-agentcore-starter-toolkit simplifies deployment
bedrock-agentcore-starter-toolkit>=1.0.0

# Strands framework - install from official source
# If package not found on PyPI, check:
# - AWS Bedrock AgentCore documentation
# - GitHub: https://github.com/aws/bedrock-agentcore-starter-toolkit
# You may need to install from GitHub or AWS package repository
strands-agents>=0.1.0

# Standard utilities
python-dotenv>=1.0.0

